appveyor: Enable 64-bit build with OpenSSL and HTTP2
Fix warnings found from that

Introduce lws_ptr_diff(head, tail) helper to normalize
pointers to char *, do the subtraction and cast the
result to int.

+++ lib/private-libwebsockets.h
@@ -721,8 +721,8 @@ enum {
 struct lws_ring {
 	void *buf;
 	void (*destroy_element)(void *element);
-	size_t buflen;
-	size_t element_len;
+	uint32_t buflen;
+	uint32_t element_len;
 	uint32_t head;
 	uint32_t oldest_tail;
 };
